moved user profile components
